A data set containing counts of the number of countries in which certain languages are recognized as (co-)official.

A long-form data set scraped from wikipedia containing an approximate number of countries in which 36 of the world's languages are recognized as being (co-)official. The data set also contains counts by continent and the country names. The table was scraped from: https://en.wikipedia.org/wiki/List_of_languages_by_the_number_of_countries_in_which_they_are_recognized_as_an_official_language

Usage

n_countries_by_language

Format

A data frame with 36 rows and 8 columns:

language

Language name in English

world

Count of contries in which language is (co-)official in the world

africa

Count of contries in which language is (co-)official in Africa

americas

Count of contries in which language is (co-)official in the Americas

asia

Count of contries in which language is (co-)official in Asia

europe

Count of contries in which language is (co-)official in Europe

oceania

Count of contries in which language is (co-)official in Oceania

countries

Vector of country names (separated by comma)

Source

https://github.com/jvcasillas/worldlanguages