A data set containing counts of the number of languages spoken in 242 countries.

A long-form data set scraped from wikipedia containing an approximate number of languages spoken in 242 countries. The data set also contains data on immigrant languages. The table was scraped from: https://en.wikipedia.org/wiki/Number_of_languages_by_country

Usage

n_languages_by_country

Format

A data frame with 242 rows and 9 columns:

rank

A numeric ranking based on number of languages spoken (lower rank = more languages)

country

Name of the country

nol_established

Number of established (i.e., non-immigrant) languages

nol_immigrant

Number of immigrant languages

nol_total

Total count of languages spoken

nol_percent

Percentage of total number of languages relative to the world, including both established and immigrant languages

nos_total

Number of total speakers

nos_mean

Mean number of total speakers per language

nos_median

Median number of total speakers per language

Source

https://github.com/jvcasillas/worldlanguages